Implemented DLL versioning on Windows. This is done using the same naming convention as on non-Windows platforms. Versioning of data files still needs to be handled.
To make it easy to link against the Juggler libraries now that the .lib file
name contains a version number, automatic linking has been implemented. This
allows builds to be simplified because they no longer have to enumerate the
Juggler .lib files and do manual matching of debug/release binaries.


git-svn-id: http://vrjuggler.googlecode.com/svn/juggler/trunk@19030 08b38cba-cd3b-11de-854e-f91c5b6e4272