The X,Y position reported for mouse events now uses a coordinate frame with the origin at the lower left-hand corner on all platforms. Previously, Win32 and X11 were using the upper left-hand corner for the origin, which is the native coordinate frame for those windowing systems. However, Juggler uses the convention of the lower left-hand corner for window positioning, so it stands to reason that mouse events would also be reported in that coordinate frame.
The most likely explanation for the discrepancy is that I screwed up when I
introduced the keyboad/mouse event queue stuff a few years back. When I added
Cocoa support, I did it correctly, which was easy anyway as Cocoa uses a lower
left-hand corner origin.

Bumped the version to 1.3.16.


git-svn-id: http://vrjuggler.googlecode.com/svn/juggler/trunk@20934 08b38cba-cd3b-11de-854e-f91c5b6e4272